requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
PyPDF2==3.0.1
reportlab==4.0.4
PyMuPDF==1.23.5
Pillow==10.0.0
Werkzeug==2.3.7
pytesseract==0.3.10
python-docx==0.8.11